requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.3
djangorestframework==3.6.3
idna==2.5
pytz==2017.2
requests==2.18.2
urllib3==1.22
